requirements.txt
pymupdf==1.21.1
pytesseract==0.3.10
pillow==9.5.0
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
transformers==4.27.4
torch==2.0.0
spacy==3.5.2
pdf2image==1.16.3
camelot-py==0.10.1
layoutparser==0.3.4